requirements.txt
fastapi
line-bot-sdk
openai
uvicorn
python-dotenv